+++ src/COBJECT/Forest.cpp
@@ -18,7 +18,7 @@ Forest::Forest(Texture2D *texture, float y, bool isSafe, float speed, int MaxObj
 }
 
 void Forest::GenerateObject() {
-    if (Objects.size() >= MaxObject) return;
+    if ((int)Objects.size() >= MaxObject) return;
     if (speed > 0 && !Objects.empty() && Objects.back()->GetX() < 100) return;
     if (speed < 0 && !Objects.empty() && Objects.back()->GetX() > screenWidth - 200) return;
     int getRand = GetRandomValue(1, 150);
